Move errors to its own file
